+++ recipes/_check_ipam_version.rb
@@ -1,5 +1,6 @@
 SUPPORTED_VERSIONS = [
-  '1.2'
+  '1.2',
+  'dev'
 ].freeze
 
 raise "Unsupported IPAM version: #{node['phpipam']['version']}" unless SUPPORTED_VERSIONS.include?(node['phpipam']['version'])